requirements.txt
# Core dependencies
pytesseract>=0.3.10
Pillow>=9.0.0
pyperclip>=1.8.0

# Cross-platform screenshot
mss>=9.0.0

# macOS native APIs for App Store sandbox compatibility
pyobjc-core>=10.0; sys_platform == 'darwin'
pyobjc-framework-Quartz>=10.0; sys_platform == 'darwin'
pyobjc-framework-Cocoa>=10.0; sys_platform == 'darwin'

# LaTeX OCR - RapidLatexOCR (lightweight, ~100MB)
rapid-latex-ocr>=0.1.0

# Windows-specific (optional)
# pywin32>=300; sys_platform == 'win32'
